refactor: getCardBrand 함수를 CardPreview 컴포넌트와 분리
